+++ cmd/hyperalert/main.go
@@ -4,39 +4,8 @@ package main
 import (
 	"os"
 
-	v "github.com/appscode/go/version"
 	"github.com/appscode/searchlight/plugins/hyperalert"
 )
-
-var (
-	Version         string
-	VersionStrategy string
-	Os              string
-	Arch            string
-	CommitHash      string
-	GitBranch       string
-	GitTag          string
-	CommitTimestamp string
-	BuildTimestamp  string
-	BuildHost       string
-	BuildHostOs     string
-	BuildHostArch   string
-)
-
-func init() {
-	v.Version.Version = Version
-	v.Version.VersionStrategy = VersionStrategy
-	v.Version.Os = Os
-	v.Version.Arch = Arch
-	v.Version.CommitHash = CommitHash
-	v.Version.GitBranch = GitBranch
-	v.Version.GitTag = GitTag
-	v.Version.CommitTimestamp = CommitTimestamp
-	v.Version.BuildTimestamp = BuildTimestamp
-	v.Version.BuildHost = BuildHost
-	v.Version.BuildHostOs = BuildHostOs
-	v.Version.BuildHostArch = BuildHostArch
-}
 
 func main() {
 	if err := hyperalert.NewCmd().Execute(); err != nil {
