+++ docs/reference/hostfacts/hostfacts.md
@@ -1,0 +1,28 @@
+## hostfacts
+
+Hostfacts by AppsCode - Node metrics
+
+### Synopsis
+
+
+Hostfacts by AppsCode - Node metrics
+
+### Options
+
+```
+      --allow_verification_with_non_compliant_keys   Allow a SignatureVerifier to use keys which are technically non-compliant with RFC6962.
+      --alsologtostderr                              log to standard error as well as files
+  -h, --help                                         help for hostfacts
+      --log_backtrace_at traceLocation               when logging hits line file:N, emit a stack trace (default :0)
+      --log_dir string                               If non-empty, write log files in this directory
+      --logtostderr                                  log to standard error instead of files (default true)
+      --stderrthreshold severity                     logs at or above this threshold go to stderr (default 2)
+  -v, --v Level                                      log level for V logs
+      --vmodule moduleSpec                           comma-separated list of pattern=N settings for file-filtered logging
+```
+
+### SEE ALSO
+* [hostfacts run](hostfacts_run.md)	 - Run operator
+* [hostfacts version](hostfacts_version.md)	 - Prints binary version number.
+
+###### Auto generated by spf13/cobra on 28-Jul-2017
