+++ docs/reference/hostfacts/hostfacts_version.md
@@ -1,0 +1,36 @@
+## hostfacts version
+
+Prints binary version number.
+
+### Synopsis
+
+
+Prints binary version number.
+
+```
+hostfacts version [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for version
+```
+
+### Options inherited from parent commands
+
+```
+      --allow_verification_with_non_compliant_keys   Allow a SignatureVerifier to use keys which are technically non-compliant with RFC6962.
+      --alsologtostderr                              log to standard error as well as files
+      --log_backtrace_at traceLocation               when logging hits line file:N, emit a stack trace (default :0)
+      --log_dir string                               If non-empty, write log files in this directory
+      --logtostderr                                  log to standard error instead of files (default true)
+      --stderrthreshold severity                     logs at or above this threshold go to stderr (default 2)
+  -v, --v Level                                      log level for V logs
+      --vmodule moduleSpec                           comma-separated list of pattern=N settings for file-filtered logging
+```
+
+### SEE ALSO
+* [hostfacts](hostfacts.md)	 - Hostfacts by AppsCode - Node metrics
+
+###### Auto generated by spf13/cobra on 28-Jul-2017
